server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "apscheduler>=3.11.1",
    "chromadb>=1.3.5",
    "croniter>=6.0.0",
    "dspy-ai>=3.0.4",
    "fastapi>=0.121.2",
    "faster-whisper>=1.2.1",
    "httpx>=0.27.0",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "pyautogui>=0.9.54",
    "pydantic-settings>=2.13.1",
    "pygetwindow>=0.0.9",
    "pypdf2>=3.0.1",
    "pyperclip>=1.11.0",
    "pyscreeze>=1.0.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.25.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "python-socketio>=5.14.3",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.6",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "sqlmodel>=0.0.27",
    "uiautomation>=2.0.29",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"

[tool.pytest.ini_options]
# pytest 配置：让 pytest 知道从当前目录（server/）导入模块
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
